Fix titles; checkpoint blogpost
